UPSTREAM: 55957: Scheduler Skips Pods that Use PVCs that Are Being Deleted
PVC Protection feature consists of the below PRs:
- https://github.com/kubernetes/kubernetes/pull/55873
- https://github.com/kubernetes/kubernetes/pull/55824
- https://github.com/kubernetes/kubernetes/pull/55957

The PRs #55873 and #55824 were merged into K8s 1.9, however, the PR #55957 was merged into K8s 1.10. That's why this PR is backported.

This commit modifies scheduler in such a way that it does not schedule a pod that uses a PVC that is being deleted.